extract script to stream logic from mchad to api
